backend/requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.4.1
billiard==3.6.4.0
black==21.7b0
celery==5.1.2
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.6
django-cors-headers==3.8.0
django-environ==0.7.0
django-filter==21.1
django-rest-framework==0.1.0
django-restql==0.15.1
djangorestframework==3.12.4
djangorestframework-filters==1.0.0.dev0
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
kombu==5.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.3.2
prompt-toolkit==3.0.20
pycodestyle==2.7.0
PyJWT==2.1.0
pyPEG2==2.15.2
pytz==2021.1
redis==3.5.3
regex==2021.8.3
six==1.16.0
sqlparse==0.4.1
tomli==1.2.1
typing-extensions==3.10.0.2
vine==5.0.0
wcwidth==0.2.5
psycopg2
dj_database_url
requests==2.26.0